Condenced functionality into Lesson3 Demo method
